requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma
langchain-google-genai
chromadb
pypdf
python-dotenv
reportlab
langchain_huggingface